Some messy changes to trainer. Added connect_4 directory for organization. nn performance still abysmal
