+++ .gitignore
@@ -1,1 +1,2 @@
 __pycache__
+logs
