+++ connect_4/connect_4_simple_trainer_report.txt
@@ -1,0 +1,16 @@
+--- Testing best connect 4 nn report at 00:24:16 ---
+	Best nn wins: 5
+	Minimax player wins: 95
+	Draws: 0
+--- Testing best connect 4 nn report at 00:33:04 ---
+	Best nn wins: 5
+	Minimax player wins: 95
+	Draws: 0
+--- Testing best connect 4 nn report at 00:41:57 ---
+	Best nn wins: 0
+	Minimax player wins: 10
+	Draws: 0
+--- Testing best connect 4 nn report at 10:09:25 ---
+	Best nn wins: 1
+	Minimax player wins: 9
+	Draws: 0
