+++ main.py
@@ -1,19 +1,35 @@
 from random_player import random_player
 from manual_player import manual_player
 from minimax_player import minimax_player
-import connect_4_value_functions
-import connect_4_nn_constructors as c4c
-import connect_4_trainers as c4t
+import connect_4.connect_4_value_functions as c4v
+import connect_4.connect_4_nn_constructors as c4c
+import connect_4.connect_4_trainers as c4t
 
 import numpy as np
 import copy
 import sys
 import datetime
 
-trainer = c4t.simple_trainer('models\\connect_4\\best_c4_simple_conv', c4c.simple_conv_connect_4_nn)
+trainer = c4t.simple_trainer('connect_4\\models\\best_c4_simple_conv', c4c.simple_conv_connect_4_nn)
 
-trainer.train(batch_size = 20, initial_load=False, rounds = 5)
+player1 = minimax_player(c4v.naive, depth=2)
+player2 = minimax_player(c4v.naive, depth=2)
 
-test_player = minimax_player(connect_4_value_functions.naive, 2)
+#trainer.generate_games('connect_4\\minimax_d2_10000_games.npy', player1=player1,player2=player2)
+"""trainer.train_from_file(
+    'connect_4\\minimax_d2_10000_games.npy',
+    initial_load=False,
+    tensorboard=True,
+    epochs=4,
+    num_test=10,
+    batch_size=32)"""
 
-trainer.test_best_nn(test_player, rounds=100)
+trainer.train2(
+    initial_load=False,
+    load_history_from='connect_4\\connect_4_nn_game_history.pkl',
+    epochs=32,
+    test_every=1,
+    rounds=16,
+    batch_size=32)
+
+trainer.test_best_nn(player1, rounds=10)
